Removed custom CSS from theme.html
The custom CSS (in the theme) is an include in the theme, pulled from
the tumblr db (it is that {CustomCSS} right above the closing style tag.
The contents of customStylesCSS (in this repository) can be copied as a
whole and pasted into the custom css textbook located in the "Advanced"
tab in the tumblr control bar.